Added DELETE
- Added DELETE budget item

+++ app/Http/Controllers/ItemManage.php
@@ -381,6 +381,7 @@ class ItemManage extends Controller
 
         return match ($item_type) {
             'allocated-expense' => $this->deleteAllocatedExpense((int) $resource_type_id, (int) $resource_id, (int) $item_id),
+            'budget' => $this->deleteBudget((int) $resource_type_id, (int) $resource_id, (int) $item_id),
             'game' => $this->deleteGame((int) $resource_type_id, (int) $resource_id, (int) $item_id),
             default => throw new \OutOfRangeException('No item type definition for ' . $item_type, 500),
         };
@@ -418,6 +419,49 @@ class ItemManage extends Controller
                 (new ItemLog())->deleteLogEntries($item_id);
                 (new ItemTransfer())->deleteTransfers($item_id);
                 (new ItemPartialTransfer())->deleteTransfers($item_id);
+                $item_type_instance->delete();
+                $item_instance->delete();
+            });
+
+            ClearCache::dispatch($cache_job_payload->payload());
+
+            return Response::successNoContent();
+        } catch (QueryException $e) {
+            return Response::foreignKeyConstraintError($e);
+        } catch (Exception $e) {
+            return Response::notFound(trans('entities.item'), $e);
+        }
+    }
+
+    private function deleteBudget(
+        string $resource_type_id,
+        string $resource_id,
+        string $item_id
+    ): JsonResponse {
+        $cache_job_payload = (new JobPayload())
+            ->setGroupKey(KeyGroup::ITEM_DELETE)
+            ->setRouteParameters([
+                'resource_type_id' => $resource_type_id,
+                'resource_id' => $resource_id
+            ])
+            ->isPermittedUser($this->hasWriteAccessToResourceType((int) $resource_type_id))
+            ->setUserId($this->user_id);
+
+        $item_model = new \App\ItemType\Budget\Models\Item();
+
+        $item_type_instance = $item_model->instance($item_id);
+        $item_instance = (new Item())->instance($resource_type_id, $resource_id, $item_id);
+
+        if ($item_instance === null || $item_type_instance === null) {
+            return Response::notFound(trans('entities.item'));
+        }
+
+        if ($item_model->hasCategoryAssignments($item_id) === true) {
+            return Response::foreignKeyConstraintCategory();
+        }
+
+        try {
+            DB::transaction(static function () use ($item_id, $item_type_instance, $item_instance) {
                 $item_type_instance->delete();
                 $item_instance->delete();
             });
